Remove phase4v3 and integrate loader
